Moved "home" to "landing"
Moved the "home" to "landing" so we can use home for a users home
planner page.
