+++ config/routes.rb
@@ -1,7 +1,7 @@
 Rails.application.routes.draw do
-  root 'pages#home'
+  root 'pages#landing'
 
-  get "pages/home"
+  get "pages/landing"
   # Define your application routes per the DSL in https://guides.rubyonrails.org/routing.html
 
   # Reveal health status on /up that returns 200 if the app boots with no exceptions, otherwise 500.
