+++ spec/requests/pages_spec.rb
@@ -1,9 +1,9 @@
 require 'rails_helper'
 
 RSpec.describe "Pages", type: :request do
-  describe "GET /home" do
+  describe "GET /landing" do
     it "returns http success" do
-      get "/pages/home"
+      get "/pages/landing"
       expect(response).to have_http_status(:success)
     end
   end
